# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Time Tracker API Tool",
    "description": "A RESTful API to help freelancers track time spent on projects and generate work summaries. Create, delete, make tasks and log time onto projects and get the summary of the project in a downloadable CSV File using this tool!",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "paths": {
...
    "/add-user": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Add User",
        "required": true,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "User details to be added",
              "required": [
                "Name",
                "Email_ID",
                "Password"
              ],
              "properties": {
                "Name": {
                  "type": "string",
                  "description": "The full name of the user.",
                  "example": "MS Dhoni"
                },
                "Email_ID": {
                  "type": "string",
                  "format": "email",
                  "description": "The email address of the user. Must be unique.",
                  "example": "thala7@gmail.com"
                },
                "Password": {
                  "type": "string",
                  "description": "The password for the user account.",
                  "example": "ThalaForAReason!"
                }
            },
                "example": {
                  "Name": "MS Dhoni",
                  "Email_ID": "thala7@gmail.com",
                  "Password": "ThalaForAReason!"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User added succesfully.",
            "headers": {
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "example": "Express"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              },
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "application/json; charset=utf-8"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "example": "54"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "W/\"36-y2NYM8DSEGW0z78RWsr8M4B6pak\""
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Thu, 02 Jan 2025 04:00:49 GMT"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "example": "timeout=5"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "message": "User Added Successfully!!",
                  "user_id": 1010
                }
              }
            }
          },
          "404": {
            "description": "Could not add user.",
            "headers": {
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "example": "Express"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              },
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "application/json; charset=utf-8"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "example": "54"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "W/\"36-y2NYM8DSEGW0z78RWsr8M4B6pak\""
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Thu, 02 Jan 2025 04:00:49 GMT"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "example": "timeout=5"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "message":"Could not add user!",
                  "error":"Duplicate User!"
                }
              }
            }
          }
        }
      }
    },
    "/verify-user": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Verify User",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "User details to get verified.",
                "required": ["user_id", "Password"],
                "properties": {
                  "user_id": {
                    "type": "number",
                    "description": "The user id of the user.",
                    "example": 1010
                  },
                  "Password": {
                    "type": "string",
                    "description": "The password for the user account.",
                    "example": "ThalaForAReason!"
                  }
                },
                "example": {
                  "user_id": 1010,
                  "Password": "ThalaForAReason!"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succesfull Login",
            "headers": {
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "example": "Express"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              },
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "application/json; charset=utf-8"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "example": "200"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "W/\"c8-G9HrPxecIQKsW+UlFce7RjgMuAs\""
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Thu, 02 Jan 2025 04:01:22 GMT"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "example": "timeout=5"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "message": "Successfully Verified User!",
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxMDEwLCJpYXQiOjE3MzU3OTA0ODIsImV4cCI6MTczNTc5NDA4Mn0.7cS1HmtU8_PuixtNF4tg1Q-W0nN-sKQIt5el0o_ZFUs"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid Credentials",
            "headers": {
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "example": "Express"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              },
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "application/json; charset=utf-8"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "example": "34"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "W/\"22-fzeLjt+7BDvY63yWDbGinFnynt4\""
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Thu, 02 Jan 2025 04:11:31 GMT"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "example": "timeout=5"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "message": "Invalid Credentials!"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "headers": {
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "example": "Express"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              },
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "application/json; charset=utf-8"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "example": "43"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "W/\"2b-3CHRB1EiLD7nhYqLlugqVtbayck\""
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Thu, 02 Jan 2025 04:13:10 GMT"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "example": "timeout=5"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "message": "No user found with ID 101011!"
                }
              }
            }
          }
        }
      }
    },
...
    "/get-project/{project_id}/tasks": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "Get Project Tasks",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "schema": {
              "type": "integer"
            },
            "description": "User details to be added",
            "required": ["project_id"],
            "properties": {
              "project_id": {
                "type": "number",
                "description": "The project id of the project.",
                "example": 1016
              },
              "example": 1016
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "X-Powered-By": {
                "schema": {
                  "type": "string",
                  "example": "Express"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string",
                  "example": "*"
                }
              },
              "Content-Type": {
                "schema": {
                  "type": "string",
                  "example": "application/json; charset=utf-8"
                }
              },
              "Content-Length": {
                "schema": {
                  "type": "integer",
                  "example": "226"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "example": "W/\"e2-QYDLWJknYl+sR59bkSP8KebH2ZQ\""
                }
              },
              "Date": {
                "schema": {
                  "type": "string",
                  "example": "Thu, 02 Jan 2025 04:04:20 GMT"
                }
              },
              "Connection": {
                "schema": {
                  "type": "string",
                  "example": "keep-alive"
                }
              },
              "Keep-Alive": {
                "schema": {
                  "type": "string",
                  "example": "timeout=5"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "task_id": 1008,
                  "project_id": 1016,
                  "user_id": 1010,
                  "task_description": "Running API Collection in Postman",
                  "status": "Ongoing",
                  "start_time": "2025-01-02 09:15:23",
                  "end_time": "2025-01-02 10:00:00",
                  "created_at": "2025-01-02 09:33:21"
                }
              }
            }
          }
        }
      }
    },
...
  }
}